% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_decennial}
\alias{census_decennial}
\title{Decennial Census}
\usage{
census_decennial(
  year = 2020,
  variables,
  for_geo,
  in_geo = NULL,
  table = "pl",
  key = NULL,
  label = FALSE
)
}
\arguments{
\item{year}{Census year (2020, 2010, 2000)}

\item{variables}{Variable names}

\item{for_geo}{Geography spec string or census_geo_*() list}

\item{in_geo}{Parent constraint}

\item{table}{Summary file: "pl" (redistricting, default), "dhc", "dp", "sf1", "sf2"}

\item{key}{API key}

\item{label}{If TRUE, add label columns}
}
\value{
tibble
}
\description{
Decennial Census
}
