% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_economic}
\alias{census_economic}
\title{Economic Census}
\usage{
census_economic(
  year = 2017,
  variables,
  for_geo,
  in_geo = NULL,
  key = NULL,
  label = FALSE
)
}
\arguments{
\item{year}{Economic Census year (2017, 2012, 2007)}

\item{variables}{Variable names (e.g. "NAICS2017", "EMP", "PAYANN")}

\item{for_geo}{Geography spec string or census_geo_*() list}

\item{in_geo}{Parent constraint (e.g. "state:06")}

\item{key}{API key}

\item{label}{If TRUE, add _label columns with variable descriptions}
}
\value{
tibble
}
\description{
Economic Census
}
