% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_flows}
\alias{census_flows}
\title{ACS Migration/Flows data}
\usage{
census_flows(year = 2022, variables, for_geo, in_geo = NULL, key = NULL)
}
\arguments{
\item{year}{ACS year}

\item{variables}{Flow variables (e.g. "MOVEDIN", "FULL1_NAME", "FULL2_NAME")}

\item{for_geo}{Geography (e.g. "county:*")}

\item{in_geo}{Parent constraint (e.g. "state:06")}

\item{key}{API key}
}
\value{
tibble
}
\description{
County-to-county or state-to-state migration flows.
}
