% Generated by build script
\name{census_geo_county}
\alias{census_geo_county}
\title{Build a county-level geography spec}
\usage{
census_geo_county()
}
\arguments{
\item{state}{State FIPS code (required)}
\item{code}{County FIPS or "*" for all (default)}
}
\value{list(for_geo, in_geo)}
\description{Build a county-level geography spec}
