% Generated by build script
\name{census_geo_place}
\alias{census_geo_place}
\title{Build a place-level geography spec}
\usage{
census_geo_place()
}
\arguments{
\item{state}{State FIPS code (required)}
\item{code}{Place FIPS or "*" for all (default)}
}
\value{list(for_geo, in_geo)}
\description{Build a place-level geography spec}
