% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_geo_state}
\alias{census_geo_state}
\title{Build a state-level geography spec}
\usage{
census_geo_state(code = "*")
}
\arguments{
\item{code}{FIPS code or "*" for all (default)}
}
\value{
list(for_geo, in_geo) for use with census_get
}
\description{
Build a state-level geography spec
}
