% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_geo_tract}
\alias{census_geo_tract}
\title{Build a tract-level geography spec}
\usage{
census_geo_tract(state, county, code = "*")
}
\arguments{
\item{state}{State FIPS code (required)}

\item{county}{County FIPS code (required)}

\item{code}{Tract code or "*" for all (default)}
}
\value{
list(for_geo, in_geo)
}
\description{
Build a tract-level geography spec
}
