% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_geo_zcta}
\alias{census_geo_zcta}
\title{Build a ZCTA geography spec}
\usage{
census_geo_zcta(code = "*")
}
\arguments{
\item{code}{ZCTA code or "*" for all (default)}
}
\value{
list(for_geo, in_geo)
}
\description{
Build a ZCTA geography spec
}
