% Generated by build script
\name{census_geography}
\alias{census_geography}
\title{List available geographies for a dataset/year}
\usage{
census_geography()
}
\arguments{
\item{dataset}{API dataset path}
\item{year}{Integer year}
\item{key}{Optional API key}
}
\value{tibble: name, geo_level, requires, wildcard}
\description{List available geographies for a dataset/year}
