% Generated by build script
\name{census_group_variables}
\alias{census_group_variables}
\title{List variables within a specific group}
\usage{
census_group_variables()
}
\arguments{
\item{dataset}{API dataset path}
\item{year}{Integer year}
\item{group}{Group name (e.g. "B01001")}
\item{key}{Optional API key}
}
\value{tibble: name, label, concept, predicate_type, group}
\description{List variables within a specific group}
