% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_variables}
\alias{census_variables}
\title{List variables for a dataset/year}
\usage{
census_variables(dataset, year, key = NULL)
}
\arguments{
\item{dataset}{API dataset path (e.g. "acs/acs5")}

\item{year}{Integer year}

\item{key}{Optional API key}
}
\value{
tibble: name, label, concept, predicate_type, group
}
\description{
List variables for a dataset/year
}
