% Generated by build script
\name{census_economic}
\alias{census_economic}
\title{Economic Census}
\usage{
census_economic()
}
\arguments{
\item{year}{Economic Census year (2017, 2012, 2007)}
\item{variables}{Variable names (e.g. "NAICS2017", "EMP", "PAYANN")}
}
\value{tibble}
\description{Economic Census}
