% Generated by build script
\name{census_flows}
\alias{census_flows}
\title{ACS Migration/Flows data}
\usage{
census_flows()
}
\arguments{
\item{year}{ACS year}
\item{variables}{Flow variables (e.g. "MOVEDIN", "FULL1_NAME", "FULL2_NAME")}
\item{for_geo}{Geography (e.g. "county:*")}
\item{in_geo}{Parent constraint (e.g. "state:06")}
\item{key}{API key}
}
\value{tibble}
\description{ACS Migration/Flows data}
