% Generated by build script
\name{census_geo_blockgroup}
\alias{census_geo_blockgroup}
\title{Build a block group geography spec}
\usage{
census_geo_blockgroup()
}
\arguments{
\item{state}{State FIPS (required)}
\item{county}{County FIPS (required)}
\item{tract}{Tract code (required)}
\item{code}{Block group code or "*" for all (default)}
}
\value{list(for_geo, in_geo)}
\description{Build a block group geography spec}
