% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_geo_county}
\alias{census_geo_county}
\title{Build a county-level geography spec}
\usage{
census_geo_county(state, code = "*")
}
\arguments{
\item{state}{State FIPS code (required)}

\item{code}{County FIPS or "*" for all (default)}
}
\value{
list(for_geo, in_geo)
}
\description{
Build a county-level geography spec
}
