% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_geo_place}
\alias{census_geo_place}
\title{Build a place-level geography spec}
\usage{
census_geo_place(state, code = "*")
}
\arguments{
\item{state}{State FIPS code (required)}

\item{code}{Place FIPS or "*" for all (default)}
}
\value{
list(for_geo, in_geo)
}
\description{
Build a place-level geography spec
}
