% Generated by build script
\name{census_geo_state}
\alias{census_geo_state}
\title{Build a state-level geography spec}
\usage{
census_geo_state()
}
\arguments{
\item{code}{FIPS code or "*" for all (default)}
}
\value{list(for_geo, in_geo) for use with census_get}
\description{Build a state-level geography spec}
