% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_geography}
\alias{census_geography}
\title{List available geographies for a dataset/year}
\usage{
census_geography(dataset, year, key = NULL)
}
\arguments{
\item{dataset}{API dataset path}

\item{year}{Integer year}

\item{key}{Optional API key}
}
\value{
tibble: name, geo_level, requires, wildcard
}
\description{
List available geographies for a dataset/year
}
