% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_groups}
\alias{census_groups}
\title{List variable groups for a dataset/year}
\usage{
census_groups(dataset, year, key = NULL)
}
\arguments{
\item{dataset}{API dataset path}

\item{year}{Integer year}

\item{key}{Optional API key}
}
\value{
tibble: name, description
}
\description{
List variable groups for a dataset/year
}
